[#3313] feat(common): refactor JsonUtils object mapper to client and service side (#3411)
### What changes were proposed in this pull request?
1. provide separate method to get object mapper in client and service
side
2. rename `JsonUtils.objectMapper` to `JsonUtils.objectMapperForTest`

### Why are the changes needed?

Fix: #3313 

### Does this PR introduce _any_ user-facing change?
no

### How was this patch tested?
existing tests
