+++ clients/client-java/src/test/java/com/datastrato/gravitino/client/TestBase.java
@@ -4,7 +4,6 @@
  */
 package com.datastrato.gravitino.client;
 
-import com.datastrato.gravitino.json.JsonUtils;
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import java.util.Collections;
@@ -22,7 +21,7 @@ import org.mockserver.model.Parameter;
 
 public abstract class TestBase {
 
-  protected static final ObjectMapper MAPPER = JsonUtils.objectMapper();
+  protected static final ObjectMapper MAPPER = ObjectMapperProvider.objectMapper();
 
   protected static ClientAndServer mockServer;
